moves favorite_button to widget folder
