+++ lib/location/location_search_delegate.dart
@@ -5,7 +5,6 @@ import 'package:flutter/material.dart';
 import 'package:latlong/latlong.dart';
 
 import 'package:trufi_app/blocs/bloc_provider.dart';
-import 'package:trufi_app/blocs/favorite_location_bloc.dart';
 import 'package:trufi_app/blocs/favorite_locations_bloc.dart';
 import 'package:trufi_app/blocs/history_locations_bloc.dart';
 import 'package:trufi_app/blocs/important_locations_bloc.dart';
@@ -15,6 +14,7 @@ import 'package:trufi_app/trufi_api.dart' as api;
 import 'package:trufi_app/trufi_localizations.dart';
 import 'package:trufi_app/trufi_models.dart';
 import 'package:trufi_app/widgets/alerts.dart';
+import 'package:trufi_app/widgets/favorite_button.dart';
 
 class LocationSearchDelegate extends SearchDelegate<TrufiLocation> {
   TrufiLocation _result;
@@ -449,87 +449,5 @@ class _SuggestionList extends StatelessWidget {
         onMapTapped(TrufiLocation.fromLatLng(description, location));
       }
     }
-  }
-}
-
-class FavoriteButton extends StatefulWidget {
-  FavoriteButton({
-    Key key,
-    this.location,
-    @required this.favoritesStream,
-  }) : super(key: key);
-
-  final TrufiLocation location;
-  final Stream<List<TrufiLocation>> favoritesStream;
-
-  @override
-  FavoriteButtonState createState() => FavoriteButtonState();
-}
-
-class FavoriteButtonState extends State<FavoriteButton> {
-  FavoriteLocationBloc _bloc;
-
-  StreamSubscription _subscription;
-
-  @override
-  void initState() {
-    super.initState();
-    _createBloc();
-  }
-
-  @override
-  void didUpdateWidget(FavoriteButton oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    _disposeBloc();
-    _createBloc();
-  }
-
-  @override
-  void dispose() {
-    _disposeBloc();
-    super.dispose();
-  }
-
-  void _createBloc() {
-    _bloc = FavoriteLocationBloc(widget.location);
-    _subscription = widget.favoritesStream.listen(_bloc.inFavorites.add);
-  }
-
-  void _disposeBloc() {
-    _subscription.cancel();
-    _bloc.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final FavoriteLocationsBloc favoriteLocationsBloc =
-        BlocProvider.of<FavoriteLocationsBloc>(context);
-    return StreamBuilder(
-      stream: _bloc.outIsFavorite,
-      builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
-        bool isFavorite = favoriteLocationsBloc.locations.contains(
-          widget.location,
-        );
-        if (isFavorite == true) {
-          return GestureDetector(
-            onTap: () {
-              favoriteLocationsBloc.inRemoveLocation.add(
-                widget.location,
-              );
-            },
-            child: Icon(Icons.favorite),
-          );
-        } else {
-          return GestureDetector(
-            onTap: () {
-              favoriteLocationsBloc.inAddLocation.add(
-                widget.location,
-              );
-            },
-            child: Icon(Icons.favorite_border),
-          );
-        }
-      },
-    );
   }
 }
